requirements.txt
annotated-types==0.7.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
confluent-kafka==2.11.1
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
fonttools==4.59.1
idna==3.10
kafka-python==2.2.15
kiwisolver==1.4.9
matplotlib==3.10.5
mysql-connector-python==9.4.0
numpy==2.3.2
oracledb==3.3.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
psycopg2==2.9.10
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
